# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ChatGPT Android Proxy",
...
    "version": "2.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8888",
      "description": "Local development"
    }
  ],
  "paths": {
...
    "/v1/files": {
      "post": {
        "summary": "Upload File",
        "description": "Sube un archivo (PDF, código, documentos) para usarlo como contexto en mensajes de chat.\n\n**Formatos soportados:** PDF, texto plano, código fuente, Markdown, DOCX, XLSX, JSON, XML, etc.\n\n**NO soportado:** Imágenes (JPEG, PNG, GIF, WebP) — el flujo anónimo no tiene capacidad de visión.\n\nEl archivo se almacena en memoria y puede referenciarse mediante su `file_id` en mensajes de chat.\n\nSi `pdfplumber` está instalado, extrae el texto de PDFs correctamente. Si no, hace un fallback básico.",
        "operationId": "upload_file",
        "tags": ["Files"],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "required": ["file"],
                "properties": {
                  "file": {
                    "type": "string",
                    "format": "binary",
                    "description": "Archivo a subir (PDF, texto, código, etc.)"
                  },
                  "purpose": {
                    "type": "string",
                    "default": "assistants",
                    "description": "Propósito del archivo (para compatibilidad con OpenAI)"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Archivo subido exitosamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {"type": "string", "example": "file-a1b2c3d4e5f6a1b2c3d4e5f6"},
                    "object": {"type": "string", "example": "file"},
                    "filename": {"type": "string", "example": "documento.pdf"},
                    "purpose": {"type": "string", "example": "assistants"},
                    "bytes": {"type": "integer", "example": 123456},
                    "created_at": {"type": "integer", "example": 1750000000},
                    "status": {"type": "string", "example": "processed"}
                  }
                }
              }
            }
          },
          "413": {
            "description": "Archivo demasiado grande (límite: 20 MB)"
          },
          "415": {
            "description": "Tipo de archivo no soportado (imágenes rechazadas)"
          },
          "422": {
            "description": "El archivo no contiene texto extraíble"
          }
        }
      },
      "get": {
        "summary": "List Files",
        "description": "Lista todos los archivos subidos en esta sesión del servidor.",
        "operationId": "list_files",
        "tags": ["Files"],
        "responses": {
          "200": {
            "description": "Lista de archivos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "object": {"type": "string", "example": "list"},
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {"type": "string"},
                          "object": {"type": "string"},
                          "filename": {"type": "string"},
                          "purpose": {"type": "string"},
                          "size": {"type": "integer"},
                          "created_at": {"type": "integer"},
                          "mime_type": {"type": "string"}
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/files/{file_id}": {
      "get": {
        "summary": "Get File",
        "description": "Obtiene los metadatos de un archivo específico (sin el contenido de texto).",
        "operationId": "get_file",
        "tags": ["Files"],
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string"},
            "example": "file-a1b2c3d4e5f6a1b2c3d4e5f6"
          }
        ],
        "responses": {
          "200": {
            "description": "Metadatos del archivo"
          },
          "404": {
            "description": "Archivo no encontrado"
          }
        }
      },
      "delete": {
        "summary": "Delete File",
        "description": "Elimina un archivo de la memoria del servidor.",
        "operationId": "delete_file",
        "tags": ["Files"],
        "parameters": [
          {
            "name": "file_id",
            "in": "path",
            "required": true,
            "schema": {"type": "string"},
            "example": "file-a1b2c3d4e5f6a1b2c3d4e5f6"
          }
        ],
        "responses": {
          "200": {
            "description": "Archivo eliminado",
            "content": {
              "application/json": {
                "example": {
                  "id": "file-a1b2c3d4e5f6a1b2c3d4e5f6",
                  "object": "file",
                  "deleted": true
                }
              }
            }
          },
          "404": {
            "description": "Archivo no encontrado"
          }
        }
      }
    },
...
  },
...
}